bank assignment
the program creates two banks in two different locations; each bank having 3 tellers and 10 customers.
the tellers are able to collect money, open and close accounts, process loan requests, provide information and issue cards.
the customers are able to make a general inquiry, deposit and withdraw money, open an account, apply for a loan and request for a card.